Add className prop to Error/Loader component/ Move and rename FeedList/FeedListInterfaces/ Add styles to FeedList component
