+++ src/actions/PostsFilter.tsx
@@ -1,4 +1,4 @@
-import { FeedPostProps } from '../FeedPost/FeedPost';
+import { FeedPostProps } from '../components/FeedPost/FeedPost';
 
 export const postFilter = (follows: string[], posts: FeedPostProps[]): FeedPostProps[] =>
   posts.filter(post => follows.includes(post.ownerUid));
