+++ src/components/Error/Error.tsx
@@ -4,11 +4,12 @@ import PropTypes from 'prop-types';
 interface ErrorProps {
   isError: boolean;
   children: React.ReactElement;
+  className?: string;
 }
 
 export const Error: React.FC<ErrorProps> = ({ isError, children }) => {
   if (isError) {
-    return <p>Error has occurred please try again...</p>;
+    return <p className="error">Error has occurred please try again...</p>;
   }
 
   return children;
@@ -16,5 +17,6 @@ export const Error: React.FC<ErrorProps> = ({ isError, children }) => {
 
 Error.propTypes = {
   isError: PropTypes.bool.isRequired,
-  children: PropTypes.element.isRequired
+  children: PropTypes.element.isRequired,
+  className: PropTypes.string
 };
