+++ src/components/FeedList/FeedList.scss
@@ -1,0 +1,27 @@
+@import '../../styles/Mixins.scss';
+@import '../../styles/Colors.scss';
+@import '../../styles/Fonts.scss';
+@import '../../styles/FontWeights.scss';
+
+.feedList {
+  @include flexColumn;
+
+  &__loader {
+    @include flexCenter;
+  }
+
+  &__posts {
+    scrollbar-width: none;
+
+    &::-webkit-scrollbar {
+      display: none;
+    }
+  }
+
+  &__endMessage {
+    font-family: $postText;
+    color: $linkedinBlue;
+    font-weight: $fontBold;
+    text-align: center;
+  }
+}
