+++ src/components/FeedList/FeedList.tsx
@@ -3,10 +3,11 @@ import InfiniteScroll from 'react-infinite-scroll-component';
 import { CircularProgress } from '@material-ui/core';
 import { useGetPosts } from '../../actions/useGetPosts';
 import { useGetUserFollows } from '../../actions/useGetUserFollows';
-import { postFilter } from './PostsFilter';
+import { postFilter } from '../../actions/PostsFilter';
 import { FeedPost, FeedPostProps } from '../FeedPost/FeedPost';
 import { Loader } from '../Loader/Loader';
 import { Error } from '../Error/Error';
+import './FeedList.scss';
 
 export const FeedList: React.FC = () => {
   const [posts, setPosts] = useState<FeedPostProps[]>([]);
@@ -15,14 +16,13 @@ export const FeedList: React.FC = () => {
 
   const { data: userFollows, isLoading: isFollowsLoading, isError: isFollowsError } = useGetUserFollows();
   const { data: allPosts, isLoading: isPostsLoading, isError: isPostsError } = useGetPosts();
+  const userPosts = postFilter(userFollows, allPosts);
 
   useEffect(() => {
     if (userFollows.length && allPosts.length) {
-      setPosts(postFilter(userFollows, allPosts).slice(0, 5));
+      setPosts(userPosts.slice(0, 5));
     }
   }, [userFollows, allPosts]);
-
-  const userPosts = postFilter(userFollows, allPosts);
 
   const getMoreData = () => {
     if (posts.length >= userPosts.length) {
@@ -33,23 +33,24 @@ export const FeedList: React.FC = () => {
     setTimeout(() => {
       setPosts(userPosts.slice(0, index + 4));
       setIndex(index + 1);
-    }, 3000);
+    }, 1000);
   };
 
   return (
-    <section>
+    <section className="feedList">
       <Loader isLoading={isFollowsLoading || isPostsLoading}>
         <Error isError={isFollowsError || isPostsError}>
           <InfiniteScroll
             dataLength={posts.length}
             next={getMoreData}
             hasMore={hasMore}
-            loader={<CircularProgress />}
-            endMessage={
-              <p style={{ textAlign: 'center' }}>
-                <b>Yay! You have seen it all</b>
+            loader={
+              <p className="feedList__loader">
+                <CircularProgress />
               </p>
             }
+            endMessage={<p className="feedList__endMessage">Yay! You have seen it all</p>}
+            className="feedList__posts"
           >
             {posts.map(({ testid, ownerUid, displayName, content, avatar, timestamp }) => (
               <FeedPost
