+++ src/components/Loader/Loader.tsx
@@ -5,11 +5,16 @@ import { CircularProgress } from '@material-ui/core';
 interface LoaderProps {
   isLoading: boolean;
   children: React.ReactElement;
+  className?: string;
 }
 
 export const Loader: React.FC<LoaderProps> = ({ isLoading, children }) => {
   if (isLoading) {
-    return <CircularProgress />;
+    return (
+      <div className="loader">
+        <CircularProgress />
+      </div>
+    );
   }
 
   return children;
@@ -17,5 +22,6 @@ export const Loader: React.FC<LoaderProps> = ({ isLoading, children }) => {
 
 Loader.propTypes = {
   isLoading: PropTypes.bool.isRequired,
-  children: PropTypes.element.isRequired
+  children: PropTypes.element.isRequired,
+  className: PropTypes.string
 };
