+++ src/shared/interfaces/StateAbleInterface.tsx
