+++ src/styles/Fonts.scss
@@ -1,0 +1,2 @@
+$postText: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans',
+  'Helvetica Neue', sans-serif;
